+++ CHANGELOG.md
@@ -3,3 +3,11 @@
 ## [1.0.0]
 
 - Initial release
+
+## [1.0.2]
+
+- Add metadata and improve marketplace
+
+## [1.0.3]
+
+- Add support for `postgres` dialect